requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.13.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==21.2.0
audioread==3.0.1
Automat==20.2.0
azfuse @ git+https://github.com/microsoft/azfuse.git@d1174016d0acd6ddda733804a48035028fefc910
azure-core==1.31.0
azure-storage-blob==12.23.1
Babel==2.8.0
banal==1.0.6
bcrypt==3.2.0
beautifulsoup4==4.12.2
bitsandbytes==0.43.3
blinker==1.4
boto3==1.35.29
botocore==1.35.29
Brotli==1.1.0
cachetools==5.3.3
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.0.3
cloud-init==23.4.4
cloud-tpu-client==0.10
cmake==3.29.3
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
config==0.5.1
configobj==5.0.6
constantly==15.1.0
contourpy==1.2.1
cryptography==3.4.8
cycler==0.12.1
Cython==3.0.10
dataset==1.6.2
dbus-python==1.2.18
debugpy==1.8.1
decorator==4.4.2
Deprecated==1.2.14
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.6.1
docker-pycreds==0.4.0
docstring_parser==0.16
dtw-python==1.5.3
einops==0.8.0
einops-exts==0.0.4
email_validator==2.2.0
emoji==2.11.1
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastdtw==0.3.4
ffmpeg==1.4
ffmpy==0.3.2
filelock==3.13.3
fire==0.6.0
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.3.1
gdown==5.1.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==1.34.1
google-api-python-client==1.8.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
h11==0.14.0
h5py==3.11.0
httpcore==1.0.6
httplib2==0.20.2
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.22.2
hydra-core==1.3.2
hyperlink==21.0.0
idna==3.3
imageio==2.35.1
imageio-ffmpeg==0.5.1
imgkit==1.2.3
importlib-metadata==4.6.4
importlib_resources==6.4.0
incremental==21.3.0
install==1.3.5
ipykernel==6.23.1
ipython==8.24.0
ipywidgets==8.1.2
isodate==0.6.1
jedi==0.19.1
jeepney==0.7.1
jieba==0.42.1
Jinja2==3.0.3
jiter==0.6.1
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.10
kaldi-io==0.9.8
keyring==23.5.0
kiwisolver==1.4.5
lapx==0.5.9
latex2mathml==3.77.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
librosa==0.10.2.post1
lightning==2.2.1
lightning-utilities==0.11.2
linkify-it-py==2.0.3
lit==18.1.6
llvmlite==0.42.0
Mako==1.3.2
markdown-it-py==2.2.0
markdown2==2.5.0
MarkupSafe==2.0.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdit-py-plugins==0.3.3
mdurl==0.1.2
more-itertools==8.10.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
mutagen==1.47.0
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.2.1
nltk==3.8.1
numba==0.59.1
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.535.133
nvidia-nccl-cu11==2.21.5
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.4.127
nvitop==1.3.2
oauth2client==4.1.3
oauthlib==3.2.0
openai==0.28.0
openai-whisper==20240930
opencv-python==4.9.0.80
orjson==3.10.5
packaging==24.0
pandas==2.2.3
parso==0.8.4
pathtools==0.1.2
pexpect==4.8.0
pillow==10.4.0
pkgconfig==1.5.5
platformdirs==4.2.1
pooch==1.8.1
proglog==0.1.10
prompt-toolkit==3.0.43
propcache==0.2.0
protobuf==5.27.3
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py-cpuinfo==9.0.0
pyarrow-hotfix==0.6
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycocoevalcap==1.2
pycocotools==2.0.8
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.7.4
pydantic_core==2.18.4
pydub==0.25.1
Pygments==2.18.0
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
PySocks==1.7.1
pysptk==1.0.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-magic==0.4.24
python-multipart==0.0.9
pytube==15.0.0
pytubefix==7.1.3
pytz==2022.1
pyworld==0.3.4
PyYAML==5.4.1
pyzmq==26.0.3
regex==2023.12.25
requests==2.32.3
responses==0.18.0
retry==0.9.2
rich==13.7.1
rouge-chinese==1.0.3
rsa==4.9
ruff==0.5.0
s3transfer==0.10.2
safetensors==0.4.2
scikit-learn==1.5.2
scipy==1.14.1
screen-resolution-extra==0.0.0
seaborn==0.13.2
SecretStorage==3.3.1
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==2.2.0
service-identity==18.1.0
setproctitle==1.3.3
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sos==4.5.6
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
SQLAlchemy==1.4.52
sqlitedict==2.1.0
sse-starlette==2.1.2
ssh-import-id==5.11
stack-data==0.6.3
starlette==0.37.2
svgwrite==1.4.3
sympy==1.13.1
systemd-python==234
tensor-parallel==1.2.4
termcolor==2.4.0
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.15.2
tomlkit==0.12.0
toolz==0.12.1
torch==2.5.0+cu118
torchaudio==2.5.0+cpu
torchvision==0.20.0+cpu
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
transformers==4.39.3
triton==3.1.0
trl @ git+https://github.com/huggingface/trl.git@bc8dfbf4e2169010b3094913a1fa4f888f750111
Twisted==22.1.0
typer==0.12.3
typing==3.7.4.3
typing_extensions==4.12.2
tyro==0.8.4
tzdata==2024.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
uc-micro-py==1.0.3
ufw==0.36.1
ujson==5.10.0
ultralytics==8.2.14
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==2.2.3
uvicorn==0.30.1
uvloop==0.19.0
wadllib==1.3.6
watchfiles==0.22.0
wavedrom==2.0.3.post3
wcwidth==0.2.13
websockets==13.1
whisper-timestamped @ git+https://github.com/linto-ai/whisper-timestamped@b6f035fb225d5ee788d478fb7db9ab81cb577800
widgetsnbextension==4.0.10
wrapt==1.16.0
xkit==0.0.0
yarl==1.16.0
yt-dlp==2024.10.22
zipp==1.0.0
zope.interface==5.4.0
